+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## TODO
 * css sourcemap
-* css ast parser remove @charset
+* css rules coverage
 
 ## Changelog